+++ src/test/java/pages/Emailp2.java
@@ -1,13 +1,13 @@
-package Pages_Sprint_2;
+package pages;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import utils.Driver;
 
-public class EmailpTest {
+public class Emailp2 {
 
-    public EmailpTest(){
+    public Emailp2(){
         PageFactory.initElements(Driver.getDriver(), this);
     }
