+++ src/test/java/pages/FlightsPage_Z.java
@@ -1,4 +1,4 @@
-package Pages;
+package pages;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
