+++ src/test/java/pages/PackagesInUrl.java
@@ -1,9 +1,7 @@
-package Pages;
+package pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
 import utils.Driver;
 
 public class PackagesInUrl {
