+++ src/test/java/pages/Paisy2.java
@@ -1,22 +1,22 @@
-package Pages;
+package pages;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import utils.Driver;
 
-public class TrueBlueButton2 {
+public class Paisy2 {
 
-    public TrueBlueButton2(){
+    public Paisy2(){
         PageFactory.initElements(Driver.getDriver(), this);
     }
 
-    @FindBy(xpath="(//span[@class='dn db-l'])[1]")
-    private WebElement trueB;
+    @FindBy(xpath="(//span[contains(text(),'Paisly')])[1]")
+    private WebElement paisy;
 
-    private String getmesage = "Sign up for TrueBlue";
-    public void enterTrue2(){
-        this.trueB.click();
+    private String getmesage = "Paisly";
+    public void enterPaisy2(){
+        this.paisy.click();
     }
 
     public String getGetmesage() {
