+++ src/test/java/pages/Travel_Insurance2.java
@@ -1,22 +1,22 @@
-package Pages;
+package pages;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import utils.Driver;
 
-public class Paisy2 {
+public class Travel_Insurance2 {
 
-    public Paisy2(){
+    public Travel_Insurance2(){
         PageFactory.initElements(Driver.getDriver(), this);
     }
 
-    @FindBy(xpath="(//span[contains(text(),'Paisly')])[1]")
-    private WebElement paisy;
+    @FindBy(xpath="(//span[contains(text(),'Travel Insurance')])[1]")
+    private WebElement traVel;
 
-    private String getmesage = "Paisly";
-    public void enterPaisy2(){
-        this.paisy.click();
+    private String getmesage = "Packages";
+    public void enterTravel2(){
+        this.traVel.click();
     }
 
     public String getGetmesage() {
