+++ src/test/java/tests/FlightsHotel.java
@@ -1,6 +1,6 @@
 package tests;
 
-import Pages.JetBlueHomePage;
+import pages.JetBlueHomePage;
 import org.openqa.selenium.Keys;
 import org.testng.Assert;
 import org.testng.annotations.Test;
