+++ src/test/java/tests/JoinTodayJoshgun2.java
@@ -1,6 +1,6 @@
 package tests;
 
-import Pages.*;
+import pages.*;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import utils.Driver;
