+++ src/test/java/tests/LoginTests_PageFactory.java
@@ -1,6 +1,6 @@
 package tests;
 
-import Pages.*;
+import pages.*;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import utils.ConfigReader;
