+++ src/test/java/tests/TestBase_For_Login_Page_Factory.java
@@ -1,9 +1,7 @@
 package tests;
 
-import Pages.SignIn;
-import Pages.SignIn2;
-import Pages.TrueBlueButton;
-import Pages.TrueBlueButton2;
+import pages.SignIn2;
+import pages.TrueBlueButton2;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.testng.Assert;
